refactor: extract ui dialog rendering
